refactor(output): replace HTMLReuslt struct with Result
