+++ .prettierignore
@@ -2,4 +2,4 @@
 **/testdata/**
 **/fixtures-go/**
 /docs/vendor/**
-/internal/output/html/*template.html
+/internal/output/html/*.gohtml
