+++ internal/output/html/package_table_template.gohtml
@@ -1,0 +1,78 @@
+<table class="vuln-table">
+  <tr>
+    <th></th>
+    <th>Package</th>
+    <th>Installed version</th>
+    <th>
+      <div class="tooltip">
+        <p>Fix available</p>
+        <span class="tooltiptext">Fix available indicates whether a patch or update exists to address the
+          vulnerabilities found in the corresponding package.</span>
+      </div>
+    </th>
+    <th>
+      <div class="tooltip">
+        <p>Vulnerability count</p>
+        <span class="tooltiptext">Vulnerability count shows the number of vulnerabilities found in the package, counted
+          by each severity level.</span>
+      </div>
+    </th>
+  </tr>
+  {{ range $rowIndex, $element := . }}
+  {{ $index := uniqueID }}
+  <tr class="table-tr package-tr {{if $element.LayerDetail }}has-layer-info"
+    data-layer="{{ $element.LayerDetail.LayerID }}{{ end }}" id="table-tr-{{ $index }}" onclick="showPackageDetails('{{ $index }}')">
+    <td class="icon-td">
+      <div class="expand-icon">
+        <i class="material-icons">play_arrow</i>
+      </div>
+    </td>
+    <td {{ if eq $element.VulnCount.AnalysisCount.Regular 0 }}class="uncalled-text" {{ end }}>{{ $element.Name }}</td>
+    <td {{ if eq $element.VulnCount.AnalysisCount.Regular 0 }}class="uncalled-text" {{ end }}>{{ $element.InstalledVersion }}</td>
+    <td>
+      {{ if ne $element.VulnCount.FixableCount.UnFixed 0 }}
+      {{ if eq $element.VulnCount.FixableCount.Fixed 0}}
+      <p class="fixable-tag no-fix">No fix available</p>
+      {{ else }}
+      <div class="tooltip">
+        <p class="fixable-tag has-fix">Partial fixes available</p>
+        <span class="tooltiptext">Upgrading the package will fix {{ $element.VulnCount.FixableCount.Fixed }} out of {{
+          $element.VulnCount.AnalysisCount.Regular }} vulnerabilities</span>
+      </div>
+      {{ end }}
+      {{ else }}
+      <p class="fixable-tag has-fix">Fix available</p>
+      {{ end }}
+    </td>
+    <td>
+      {{ if ne $element.VulnCount.AnalysisCount.Regular 0 }}
+      {{ template "severity_summary_template.gohtml" $element.VulnCount.SeverityCount }}
+      {{ else }}
+      <p class="uncalled-text">{{ getFilteredVulnReasons $element.HiddenVulns }}</p>
+      {{ end }}
+    </td>
+  </tr>
+  <tr>
+    <td colspan="100%">
+      <div id="table-tr-{{ $index }}-details" class="package-details">
+        {{ if and $element.LayerDetail (not (eq $element.LayerDetail.LayerID "")) }}
+        <div class="layer-command-container">
+          <span class="package-detail-title">Layer introduced in:</span>
+          {{ if ne $element.LayerDetail.LayerCommandDetailed "" }}
+          <div class="tooltip">{{ $element.LayerDetail.LayerCommand }}
+            <span class="tooltiptext">{{ $element.LayerDetail.LayerCommandDetailed }}</span>
+          </div>
+          {{ else }}
+          <div>{{ $element.LayerDetail.LayerCommand }}</div>
+          {{ end }}
+        </div>
+        <p><span class="package-detail-title">Layer ID:</span> {{ $element.LayerDetail.LayerID }}</p>
+        <p><span class="package-detail-title">In base image:</span> {{ $element.LayerDetail.InBaseImage }}</p>
+        {{ end }}
+        {{ template "vuln_table_template.gohtml" $element }}
+      </div>
+
+    </td>
+  </tr>
+  {{ end }}
+</table>
