+++ internal/output/html/package_view_template.gohtml
@@ -1,11 +1,11 @@
 {{ range . }}
 <div class="ecosystem-container{{ if .IsOS }} os-type{{ else }} project-type{{ end }}">
-  <h2 class="ecosystem-heading">{{ .Ecosystem }}</h2>
+  <h2 class="ecosystem-heading">{{ .Name }}</h2>
   <div class="ecosystem-sources-container">
     {{ range .Sources }}
     <div class="source-container">
-      <h3 class="source-heading">Source: {{ .Source }}</h3>
-      {{ template "package_table_template.html" .PackageResults }}
+      <h3 class="source-heading">Source: {{ .Name }}</h3>
+      {{ template "package_table_template.gohtml" .Packages }}
     </div>
     {{ end }}
   </div>
