+++ internal/output/html/severity_summary_template.gohtml
