+++ internal/output/html/vuln_table_template.gohtml
@@ -1,0 +1,86 @@
+<table class="inner-table">
+  <tr>
+    <th>Vulnerability ID</th>
+    <th>Aliases</th>
+    <th>Fixed version</th>
+    <th class="severity-cell">Severity</th>
+    <th class="open-in-tab-cell"></th>
+  </tr>
+  {{ range $rowIndex, $element := .RegularVulns }}
+  {{ $index := uniqueID }}
+  <tr class='table-tr vuln-tr' id="table-tr-{{ $index }}" data-vuln-id="{{ $element.ID }}">
+    <td>
+      {{ if eq (len $element.GroupIDs) 0 }}
+      {{ $element.ID }}
+      {{ else }}
+      <div class="tooltip">
+        {{ $element.ID }}
+        <span class="tooltiptext">Group IDs: {{ join $element.GroupIDs ", " }}</span>
+      </div>
+      {{ end }}
+    </td>
+    <td>
+      {{ if eq (len $element.Aliases) 1 }}
+      {{ index $element.Aliases 0 }}
+      {{ else if gt (len $element.Aliases) 1}}
+      <div class="tooltip">
+        {{ index $element.Aliases 0}}, ...
+        <span class="tooltiptext">{{ join $element.Aliases ", " }}</span>
+      </div>
+      {{ end }}
+    </td>
+
+    <td>
+      <p {{ if not $element.IsFixable }} class="fixable-tag no-fix" {{ end }}>
+        {{$element.FixedVersion }}</p>
+    </td>
+    <td class="severity-cell">
+      <div id="{{ formatRating $element.SeverityRating }}-short" class="severity-short">
+        <p class="{{ formatRating $element.SeverityRating }}">{{ $element.SeverityScore }}</p>
+      </div>
+    </td>
+    <td class="open-in-tab-cell">
+      <p class="open-in-tab-tag" onclick="openVulnInNewTab('{{ $element.ID }}')">Open in tab</p>
+    </td>
+  </tr>
+  {{ end }}
+
+  {{ range $rowIndex, $element := .HiddenVulns }}
+  {{ $index := uniqueID }}
+  <tr class='table-tr vuln-tr uncalled-tr' id="table-tr-{{ $index }}" data-vuln-id="{{ $element.ID }}">
+    <td class="uncalled-text">
+      {{ if eq (len $element.GroupIDs) 0 }}
+      {{ $element.ID }}
+      {{ else }}
+      <div class="tooltip">
+        {{ $element.ID }}
+        <span class="tooltiptext">Group IDs: {{ join $element.GroupIDs ", " }}</span>
+      </div>
+      {{ end }}
+    </td>
+    <td class="uncalled-text">
+      {{ if eq (len $element.Aliases) 1 }}
+      {{ index $element.Aliases 0 }}
+      {{ else if gt (len $element.Aliases) 1}}
+      <div class="tooltip">
+        {{ index $element.Aliases 0}}...
+        <span class="tooltiptext">{{ join $element.Aliases ", " }}</span>
+      </div>
+      {{ end }}
+    </td>
+
+    <td>
+      <p {{ if not $element.IsFixable }} class="fixable-tag no-fix" {{ else }}
+        class="uncalled-text" {{ end }}>{{$element.FixedVersion }}</p>
+    </td>
+    <td class="severity-cell">
+      <div id="{{ formatRating $element.SeverityRating }}-short" class="severity-short">
+        <p class="{{ formatRating $element.SeverityRating }}">{{ $element.SeverityScore }}</p>
+      </div>
+    </td>
+    <td class="open-in-tab-cell">
+      <p class="open-in-tab-tag" onclick="openVulnInNewTab('{{ $element.ID }}')">Open in tab</p>
+    </td>
+  </tr>
+  {{ end }}
+</table>
