--- internal/output/html/vuln_table_template.html
@@ -1,86 +1,0 @@
-<table class="inner-table">
-  <tr>
-    <th>Vulnerability ID</th>
-    <th>Aliases</th>
-    <th>Fixed version</th>
-    <th class="severity-cell">Severity</th>
-    <th class="open-in-tab-cell"></th>
-  </tr>
-  {{ range $rowIndex, $element := .CalledVulns }}
-  {{ $index := uniqueID }}
-  <tr class='table-tr vuln-tr' id="table-tr-{{ $index }}" data-vuln-id="{{ $element.Summary.ID }}">
-    <td>
-      {{ if eq (len $element.Detail.GroupIDs) 0 }}
-      {{ $element.Summary.ID }}
-      {{ else }}
-      <div class="tooltip">
-        {{ $element.Summary.ID }}
-        <span class="tooltiptext">Group IDs: {{ join $element.Detail.GroupIDs ", " }}</span>
-      </div>
-      {{ end }}
-    </td>
-    <td>
-      {{ if eq (len $element.Detail.Aliases) 1 }}
-      {{ index $element.Detail.Aliases 0 }}
-      {{ else if gt (len $element.Detail.Aliases) 1}}
-      <div class="tooltip">
-        {{ index $element.Detail.Aliases 0}}, ...
-        <span class="tooltiptext">{{ join $element.Detail.Aliases ", " }}</span>
-      </div>
-      {{ end }}
-    </td>
-
-    <td>
-      <p {{ if eq $element.Summary.FixedVersion "No fix available" }} class="fixable-tag no-fix" {{ end }}>
-        {{$element.Summary.FixedVersion }}</p>
-    </td>
-    <td class="severity-cell">
-      <div id="{{ toLower $element.Summary.SeverityRating }}-short" class="severity-short">
-        <p class="{{ toLower $element.Summary.SeverityRating }}">{{ $element.Summary.SeverityScore }}</p>
-      </div>
-    </td>
-    <td class="open-in-tab-cell">
-      <p class="open-in-tab-tag" onclick="openVulnInNewTab('{{ $element.Summary.ID }}')">Open in tab</p>
-    </td>
-  </tr>
-  {{ end }}
-
-  {{ range $rowIndex, $element := .UncalledVulns }}
-  {{ $index := uniqueID }}
-  <tr class='table-tr vuln-tr uncalled-tr' id="table-tr-{{ $index }}" data-vuln-id="{{ $element.Summary.ID }}">
-    <td class="uncalled-text">
-      {{ if eq (len $element.Detail.GroupIDs) 0 }}
-      {{ $element.Summary.ID }}
-      {{ else }}
-      <div class="tooltip">
-        {{ $element.Summary.ID }}
-        <span class="tooltiptext">Group IDs: {{ join $element.Detail.GroupIDs ", " }}</span>
-      </div>
-      {{ end }}
-    </td>
-    <td class="uncalled-text">
-      {{ if eq (len $element.Detail.Aliases) 1 }}
-      {{ index $element.Detail.Aliases 0 }}
-      {{ else if gt (len $element.Detail.Aliases) 1}}
-      <div class="tooltip">
-        {{ index $element.Detail.Aliases 0}}...
-        <span class="tooltiptext">{{ join $element.Detail.Aliases ", " }}</span>
-      </div>
-      {{ end }}
-    </td>
-
-    <td>
-      <p {{ if eq $element.Summary.FixedVersion "No fix available" }} class="fixable-tag no-fix" {{ else }}
-        class="uncalled-text" {{ end }}>{{$element.Summary.FixedVersion }}</p>
-    </td>
-    <td class="severity-cell">
-      <div id="{{ toLower $element.Summary.SeverityRating }}-short" class="severity-short">
-        <p class="{{ toLower $element.Summary.SeverityRating }}">{{ $element.Summary.SeverityScore }}</p>
-      </div>
-    </td>
-    <td class="open-in-tab-cell">
-      <p class="open-in-tab-tag" onclick="openVulnInNewTab('{{ $element.Summary.ID }}')">Open in tab</p>
-    </td>
-  </tr>
-  {{ end }}
-</table>
